requirements.txt
streamlit
openai
pydantic
python-dotenv
pandas
numpy
scikit-learn
matplotlib
soundfile